Move SchemaComponentUpdate to submodule
